+++ immutable-arrays/core/src/main/kotlin/com/danrusu/pods4k/immutableArrays/ImmutableIntArray.kt
@@ -225,6 +225,11 @@ public value class ImmutableIntArray @PublishedApi internal constructor(
     public inline fun any(predicate: (predicate: Int) -> Boolean): Boolean = values.any(predicate)
 
     /**
+     * See [IntArray.none]
+     */
+    public inline fun none(predicate: (predicate: Int) -> Boolean): Boolean = values.none(predicate)
+
+    /**
      * Leaves this immutable array as is and returns an ImmutableIntArray with all elements sorted
      * according to the natural sort order of the value returned by the [selector].
      *