+++ immutable-arrays/core/src/main/kotlin/com/danrusu/pods4k/immutableArrays/ImmutableShortArray.kt
@@ -227,6 +227,12 @@ public value class ImmutableShortArray @PublishedApi internal constructor(
     public inline fun any(predicate: (predicate: Short) -> Boolean): Boolean = values.any(predicate)
 
     /**
+     * See [ShortArray.none]
+     */
+    public inline fun none(predicate: (predicate: Short) -> Boolean): Boolean =
+            values.none(predicate)
+
+    /**
      * Leaves this immutable array as is and returns an ImmutableShortArray with all elements sorted
      * according to the natural sort order of the value returned by the [selector].
      *