+++ changelog.md
@@ -9,8 +9,9 @@ _Release Date TBD_
 
 **New Features:**
 
-* Add `all(predicate)` method
+* Add `none(predicate)` method
 * Add `any(predicate)` method
+* Add `all(predicate)` method
 
 ## 0.2.0